Add messages service.
Inject it in the hero service (service-in-service) and use it to add a
message when we fetch the list of heroes with getHeroes().

Inject it in the heroes component and use it to add a message when a
hero is selected.

Inject it as a public property in the messages component so that --
besides displaying the list of messages --- messageService.clear() is
called when a button in the template is clicked.